articles/identifying-diagnostic-regions-in-viruses/code/pyproject.toml
[project]
name = "identify"
version = "0.1.0"
description = ""
authors = [
    {name = "Lachlan Miller",email = "lachlan@lachlan-miller.me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "pyqt6 (>=6.8.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
